+++ CHANGELOG.md
@@ -11,6 +11,8 @@ for further information about branching and tagging conventions.
 #### Added
 - Atomic swap requests history in atomic swap details
 - Update asset form
+- Set default quote asset for business in settings
+- Convert asset for each business
 
 #### Deleted
 - Dropdown account balances